requirements.txt
discord.py
ollama
requests
python-dotenv
